requirements.txt
PyPDF2>=3.0
python-docx>=1.0
openpyxl>=3.1
pandas>=2.0
ipywidgets>=8.0
